requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
typing_extensions==4.11.0
yarl==1.9.4
